Automatically Suspend Context If Unused
Added an optional flag to allow the audio context to suspend itself if unused after a delay. When enabled this fixes a bug in certain mobile devices running chrome (Samsung Galaxy Tab 2,3,4) where the context is silently disabled by the browser if it sits unused for too long. It also should reduce power consumption.